change the location of extending pools methods to be in tf-grid-soultion instead of threebot_deployer and adding mail service
