+++ jumpscale/packages/tfgrid_solutions/services/auto_extend_pools.py
@@ -1,0 +1,17 @@
+from jumpscale.tools.servicemanager.servicemanager import BackgroundService
+from jumpscale.packages.tfgrid_solutions.utils import auto_extend_pools
+
+
+class AutoExtendPoolService(BackgroundService):
+    def __init__(self, name="threebot_deployer_auto_extend_pools", interval=60 * 60 * 24, *args, **kwargs):
+        """
+        Test service that runs every 1 day
+        """
+        super().__init__(name, interval, *args, **kwargs)
+
+    def job(self):
+        auto_extend_pools()
+        print("Auto extend pool service done successfully")
+
+
+service = AutoExtendPoolService()
